Add files via upload
You are provided with a template file ProjectB.py, POSCAR_AlNTi2 (Figure 2), and PARTCAR_SSeZn (Figure 6). The template file ProjectB.py is for you to modify using instructions provided in the following sections. No external packages are needed. Materials Science and Engineering is a field that comb. Crystals are ordered materials characterized by a repeating arrangement of atoms within a unit cell, analogous to the bricks in a wall, where atoms sit at specific positions and can be substituted with other elements to create new materials. This structure is defined by a basis set of lattice vectors and atomic positions, essential for designing novel materials through computational methods by altering atomic sites, as exemplified by the potential substitution of nitrogen with carbon in a unit cell